university_storaging/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2026.3
cron-descriptor==1.4.5
dj-config-url==0.1.1
Django==6.0.3
django-celery-beat==2.9.0
django-crispy-forms==2.6
django-environ==0.13.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==7.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
